feat: lock batch implementation, sort batches by name by default on home page

+++ src/Kraken/Controllers/Api/ReleaseBatchesController.cs
@@ -3,6 +3,7 @@ namespace Kraken.Controllers.Api
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Net;
     using System.Threading.Tasks;
     using Kraken.Filters;
     using Kraken.Models;
@@ -11,6 +12,7 @@ namespace Kraken.Controllers.Api
     using Microsoft.AspNet.Http;
     using Microsoft.AspNet.Mvc;
     using Microsoft.Data.Entity;
+    using Octopus.Client.Exceptions;
     using Octopus.Client.Model;
 
     [Authorize]
@@ -31,25 +33,9 @@ namespace Kraken.Controllers.Api
 
         // GET: api/ReleaseBatches
         [HttpGet]
-        public IEnumerable<ReleaseBatch> GetReleaseBatches([FromQuery] string orderBy)
+        public IEnumerable<ReleaseBatch> GetReleaseBatches()
         {
-            Func<ReleaseBatch, object> orderByFunc;
-            switch (orderBy)
-            {
-                case "Name":
-                    orderByFunc = batch => batch.Name;
-                    break;
-                case "Updated":
-                    orderByFunc = batch => batch.UpdateDateTime;
-                    break;
-                case "Id":
-                    orderByFunc = batch => batch.Id;
-                    break;
-                default:
-                    orderByFunc = batch => batch.Id;
-                    break;
-            }
-            return _context.ReleaseBatches.OrderBy(orderByFunc);
+            return _context.ReleaseBatches.OrderBy(rb => rb.Name);
         }
 
         // GET: api/ReleaseBatches/5
@@ -109,6 +95,11 @@ namespace Kraken.Controllers.Api
                 return HttpNotFound();
             }
 
+            if (existingReleaseBatch.IsLocked)
+            {
+                return GetLockedForbiddenUpdateResult(existingReleaseBatch);
+            }
+
             if (releaseBatch.Name != null)
             {
                 existingReleaseBatch.Name = releaseBatch.Name;
@@ -124,6 +115,32 @@ namespace Kraken.Controllers.Api
             await _context.SaveChangesAsync();
 
             return new HttpStatusCodeResult(StatusCodes.Status204NoContent);
+        }
+
+        [HttpPut("{idOrName}/LockReleaseBatch")]
+        public async Task<IActionResult> LockReleaseBatch([FromRoute] string idOrName, [FromBody] string comment)
+        {
+            if (!ModelState.IsValid)
+            {
+                return HttpBadRequest(ModelState);
+            }
+
+            var releaseBatch = await GetReleaseBatch(idOrName, false, false);
+            releaseBatch.IsLocked = !releaseBatch.IsLocked;
+            if (releaseBatch.IsLocked)
+            {
+                releaseBatch.LockComment = comment;
+                releaseBatch.LockUserName = User.Identity.Name;
+            }
+            else
+            {
+                releaseBatch.LockComment = string.Empty;
+                releaseBatch.LockUserName = string.Empty;
+            }
+
+            await _context.SaveChangesAsync();
+
+            return Ok(releaseBatch.IsLocked);
         }
 
         // PUT: api/ReleaseBatches/5/Logo
@@ -217,6 +234,11 @@ namespace Kraken.Controllers.Api
                 return HttpNotFound();
             }
 
+            if (releaseBatch.IsLocked)
+            {
+                return GetLockedForbiddenUpdateResult(releaseBatch);
+            }
+
             _context.ReleaseBatches.Remove(releaseBatch);
             await _context.SaveChangesAsync();
 
@@ -236,6 +258,11 @@ namespace Kraken.Controllers.Api
             if (releaseBatch == null)
             {
                 return HttpNotFound();
+            }
+
+            if (releaseBatch.IsLocked)
+            {
+                return GetLockedForbiddenUpdateResult(releaseBatch);
             }
 
             var projectResource = _octopusProxy.GetProject(requestBody.ProjectIdOrSlugOrName);
@@ -297,6 +324,11 @@ namespace Kraken.Controllers.Api
                 return HttpNotFound();
             }
 
+            if (releaseBatch.IsLocked)
+            {
+                return GetLockedForbiddenUpdateResult(releaseBatch);
+            }
+
             var projectResource = _octopusProxy.GetProject(projectIdOrSlugOrName);
             if (projectResource == null)
             {
@@ -334,6 +366,11 @@ namespace Kraken.Controllers.Api
             if (releaseBatch == null)
             {
                 return HttpNotFound();
+            }
+
+            if (releaseBatch.IsLocked)
+            {
+                return GetLockedForbiddenUpdateResult(releaseBatch);
             }
 
             if (releaseBatch.Items != null && releaseBatch.Items.Any())
@@ -391,13 +428,25 @@ namespace Kraken.Controllers.Api
                 return HttpBadRequest("Environment Not Found");
             }
 
-            var deployments = new List<DeploymentResource>();
+            var responseBody = new DeployBatchResponseBody
+            {
+                SuccessfulProjects = new List<string>(),
+                FailedProjects = new List<string>()
+            };
 
             if (releaseBatch.Items != null && releaseBatch.Items.Any())
             {
                 foreach (var releaseBatchItem in releaseBatch.Items.Where(releaseBatchItem => !string.IsNullOrEmpty(releaseBatchItem.ReleaseId)))
                 {
-                    deployments.Add(_octopusProxy.DeployRelease(releaseBatchItem.ReleaseId, environment.Id, forceRedeploy));
+                    try
+                    {
+                        _octopusProxy.DeployRelease(releaseBatchItem.ReleaseId, environment.Id, forceRedeploy);
+                        responseBody.SuccessfulProjects.Add(releaseBatchItem.ProjectName);
+                    }
+                    catch (OctopusSecurityException)
+                    {
+                        responseBody.FailedProjects.Add(releaseBatchItem.ProjectName);
+                    }
                 }
             }
 
@@ -408,7 +457,12 @@ namespace Kraken.Controllers.Api
 
             await _context.SaveChangesAsync();
 
-            return Ok(deployments);
+            if (!responseBody.SuccessfulProjects.Any())
+            {
+                return new HttpStatusCodeResult(StatusCodes.Status403Forbidden);
+            }
+
+            return Ok(responseBody);
         }
 
         // GET: api/ReleaseBatches/5/PreviewReleases
@@ -536,12 +590,12 @@ namespace Kraken.Controllers.Api
                 {
                     Id = e.Id,
                     Name = e.Name
-                }),
+                }).ToList(),
                 Deploy = environmentsWithPermissions[Permission.DeploymentCreate].Select(e => new EnvironmentMapping
                 {
                     Id = e.Id,
                     Name = e.Name
-                }),
+                }).ToList(),
             };
 
             return Ok(retVal);
@@ -578,12 +632,16 @@ namespace Kraken.Controllers.Api
                 releaseBatch = await query.SingleOrDefaultAsync(b => b.Id == id);
             }
 
-            if (releaseBatch == null)
-            {
-                releaseBatch = await query.SingleOrDefaultAsync(b => b.Name == idOrName);
-            }
+            return releaseBatch ?? await query.SingleOrDefaultAsync(b => b.Name == idOrName);
+        }
 
-            return releaseBatch;
+        private ObjectResult GetLockedForbiddenUpdateResult(ReleaseBatch releaseBatch)
+        {
+            var lockedReason = $"Locked by: {releaseBatch.LockUserName}{(!string.IsNullOrEmpty(releaseBatch.LockComment) ? $" with comment: {releaseBatch.LockComment}" : "")}";
+            return new ObjectResult(lockedReason)
+            {
+                StatusCode = StatusCodes.Status403Forbidden
+            };
         }
 
         private readonly ApplicationDbContext _context;
@@ -594,6 +652,12 @@ namespace Kraken.Controllers.Api
         {
             public string ProjectIdOrSlugOrName { get; set; }
             public string ReleaseVersion { get; set; }
+        }
+
+        public class DeployBatchResponseBody
+        {
+            public List<string> SuccessfulProjects { get; set; } 
+            public List<string> FailedProjects { get; set; } 
         }
 
         public class ProjectProgressResponseBody
@@ -609,8 +673,8 @@ namespace Kraken.Controllers.Api
 
         public class EnvironmentsWithPermissionsResponseBody
         {
-            public IEnumerable<EnvironmentMapping> View { get; set; }
-            public IEnumerable<EnvironmentMapping> Deploy { get; set; }
+            public List<EnvironmentMapping> View { get; set; }
+            public List<EnvironmentMapping> Deploy { get; set; }
         }
 
         // json serializer was having issues with returning EnvironmentsWithPermissionsResponseBody as View and Deploy point to the same objects
